requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
alembic==1.13.2
authlib==1.3.1
httpx==0.27.2
itsdangerous==2.2.0
